# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  entities.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/entities.User'
    type: object
  entities.CreateDefectRequest:
    properties:
      external_link:
        type: string
      result_id:
        type: string
      severity:
        type: string
      title:
        type: string
    type: object
  entities.CreateOrganizationMemberRequest:
    properties:
      role:
        type: string
      user_id:
        type: string
    type: object
  entities.CreateOrganizationRequest:
    properties:
      name:
        type: string
    type: object
  entities.CreateProjectRequest:
    properties:
      created_by:
        type: string
      description:
        type: string
      name:
        type: string
      org_id:
        type: string
    type: object
...
  entities.Defect:
    properties:
      created_at:
        type: string
      external_link:
        type: string
      id:
        type: string
      result_id:
        type: string
      severity:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  entities.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  entities.Organization:
    properties:
      created_at:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  entities.OrganizationMember:
    properties:
      org_id:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
  entities.Project:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      org_id:
        type: string
    type: object
  entities.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  entities.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  title: Plain-TCMS API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Log in a user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entities.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /defects/{id}:
    delete:
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete defect
      tags:
      - Defects
    get:
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Defect'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get defect by ID
      tags:
      - Defects
    put:
      consumes:
      - application/json
      parameters:
      - description: Defect ID
        in: path
        name: id
        required: true
        type: string
      - description: Defect
        in: body
        name: defect
        required: true
        schema:
          $ref: '#/definitions/entities.CreateDefectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Defect'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update defect
      tags:
      - Defects
  /organizations:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Organization'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get all organizations
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization
        in: body
        name: organization
        required: true
        schema:
          $ref: '#/definitions/entities.CreateOrganizationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Organization'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create organization
      tags:
      - Organizations
  /organizations/{id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete organization
      tags:
      - Organizations
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Organization'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get organization by ID
      tags:
      - Organizations
  /organizations/{id}/members:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.OrganizationMember'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get organization members
      tags:
      - Organizations
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: Member
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/entities.CreateOrganizationMemberRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.OrganizationMember'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Add member to organization
      tags:
      - Organizations
  /organizations/{id}/members/{user_id}:
    delete:
      parameters:
      - description: Organization ID
        in: path
        name: id
        required: true
        type: string
      - description: User ID
        in: path
        name: user_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Remove member from organization
      tags:
      - Organizations
  /organizations/{org_id}/projects:
    get:
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get projects by organization
      tags:
      - Projects
    post:
      consumes:
      - application/json
      parameters:
      - description: Organization ID
        in: path
        name: org_id
        required: true
        type: string
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/entities.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/entities.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create project
      tags:
      - Projects
  /projects:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entities.Project'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get all projects
      tags:
      - Projects
  /projects/{id}:
    delete:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete project
      tags:
      - Projects
    get:
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Project'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get project by ID
      tags:
      - Projects
    put:
      consumes:
      - application/json
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/entities.CreateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.Project'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Update project
      tags:
      - Projects
...
swagger: "2.0"